% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allqs}
\alias{allqs}
\title{Parsed-out Reformatted Priors from Round Two of Expert Elicitation of Right Whale Movement - June 2015}
\format{A recursive list with 6 elements, each of which is an 8 element list}
\usage{
allqs
}
\description{
A recursive list that holds the reformatted priors from each individual
expert for each question. There are 6 elements in the list, corresponding
to each of the 6 questions. Within each of these 6 elements, there are
8 elements, corresponding to each of the 8 experts. So there are 48
elements in total that we'll use to choose priors as we iterate through
the gibbs loop, i.e. they are the inputs to calculating the c coefficients
and then the K coefficients. We then use the Kth prior each time through
the loop.
}
\keyword{datasets}

