% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcC.R
\name{calcC}
\alias{calcC}
\title{Calculate c values for use in the choosing of an experts' prior}
\usage{
calcC(data, prior)
}
\description{
This function calculates c weights that are the first step in
choosing an individual expert's prior prior to sampling
from the Dirichlet distribution within the Gibbs loop
}

