% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcC.R
\name{calcK}
\alias{calcK}
\title{Calculate k probability vector for use in choosing of an experts' prior}
\usage{
calcK(cdat)
}
\description{
The goal of this function is to pull in all the output from \code{calcC}
to return a K probability vector. This vector will then be used in
conjunction with \code{rmultinom()} to choose an expert's prior, $k_i$.
}

