% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleveDot.R
\name{cleveDot}
\alias{cleveDot}
\title{cleveDot - Creates faceted Cleveland-style strip/dot charts of Movement Probabilities}
\usage{
cleveDot(n, probs, from, to)
}
\arguments{
\item{n}{A number corresponding to the number of geographic regions}

\item{probs}{Posterior movement probabilities}

\item{from}{A label of length 1 corresponding to the from region}

\item{to}{A vector of length n corresponding to the to regions}
}
\description{
cleveDot - Creates faceted Cleveland-style strip/dot charts of Movement Probabilities
}
\examples{
cleveDot(n = 9, probs = c(0.033, 0.033, 0.033, 0.184, 0.033, 0.584, 0.033, 0.033, 0.033), from = 'BOF',
to = c('BOF', 'GOM', 'GSC', 'JL', 'MIDA', 'NE', 'NRTH', 'RB', 'SEUS'))
}
\seealso{
I used recipe 3.10 from The R Graphics Cookbook, by Winston Chang
}

