% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfList}
\alias{dfList}
\title{Answers from Round Two of Expert Elicitation on Right Whale Movement - June 2015}
\format{A list with 6 elements, corresponding to 6 questions. Variables are:
\itemize{
\item inputConfRaw: the raw value the expert entered for their confidence around their answers (1-100)
\item inputConfScale: the response for inputConfRaw rescaled range (1-622)
\item moveNortherntoMida: Response for number of whales out of 100 moving from Northern regions to mid-Atlantic. note that the variable number changes from 2 to 3 for questions 2 and 5.
\item gender: the sex of the animal
\item expert: the expert's numerical identifier
\itme question: the number of each question ("q9round2", "q10round2", "q11round2", "q12round2", "q13round2", "q14round2")
}}
\usage{
dfList
}
\description{
A dataset in list format containing the answers from 8 different experts on 6 different
questions pertaining to movements of right whales in and around the mid-Atlantic
region. Three questions were on Adult male right whales; 3 questions were on Adult
female right whales. List elements correspond to each question.
}
\seealso{
We used a list because the number of variables changes in each question.
}
\keyword{datasets}

