% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flatFem}
\alias{flatFem}
\title{Posterior Movement Probabilities for Females (Flat Prior)}
\format{An array 9 rows, 9 columns and 12 slices}
\usage{
flatFem
}
\description{
A dataset containing the estimated movement probabilities from
one region (columns) to another region (rows). Each of 12 slices
of the array corresponds to a month. Slice 1 corresponds to
estimated movements from January to February, etc. These were
created using a flat prior:
Dir(1, 1, 1, 1, 1, 1, 1, 1, 1)
}
\keyword{datasets}

