% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{june2015}
\alias{june2015}
\title{Answers from Second Expert Elicitation on Right Whale Movement - June 2015}
\format{A data frame with 48 rows and 7 variables:
\itemize{
\item inputConfRaw: the raw value the expert entered for their confidence around their answers (1-100)
\item inputConfScale: the response for inputConfRaw rescaled range (1-622)
\item moveMIDA: Response for number of whales out of 100 moving into/out of the MIDA.
\item gender: the sex of the animal
\item expert: Numerical identifier of the expert (1-10)
\item question: the number of each question ("q9round2", "q10round2", "q11round2", "q12round2", "q13round2", "q14round2")
}}
\usage{
june2015
}
\description{
A dataset containing the summarised answers from 8 different experts on 6 different
questions pertaining to the movements of right whales in and around the mid-Atlantic
region (Questions 9-14).
}
\keyword{datasets}

