% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makebdata.R, R/maketridata.R
\name{makebdata}
\alias{makebdata}
\title{makebdata - Makes a long form of the data for plotting}
\usage{
makebdata(nov2013)

makebdata(nov2013)
}
\description{
makebdata - Makes a long form of the data for plotting

maketridata - Makes a long form of the data from november 2015 for plotting
}
\examples{
makebdata(june2015)
makebdata(nov2013)
}

