% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeConn.R
\name{makeConn}
\alias{makeConn}
\title{makeConn - Creates curved connections between Habitat Regions}
\usage{
makeConn(n, coords, probs)
}
\arguments{
\item{n}{A number corresponding to the number of geographic regions}

\item{coords}{A set of x,y coordinates of nrow == n}
}
\description{
makeConn - Creates curved connections between Habitat Regions
}
\examples{
makeConn(3, matrix(dat = c(1, 4, 7, 1, 5, 6), nrow = 3, ncol = 2), c(.4, .5, .1))
}

