% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePriorMat.R
\name{makePriorMat}
\alias{makePriorMat}
\title{makePriorMat - list elements with values from the elicitation and converts them to prior matrices}
\usage{
makePriorMat()
}
\arguments{
\item{list1}{A list element}
}
\description{
\code{makePriorMat} is used to
}

