% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nov2013}
\alias{nov2013}
\title{Answers from Round One of Expert Elicitation on Right Whale Movement - November 2013}
\format{A data frame with 140 rows and 8 variables:
\itemize{
\item question: the number of each question (1-14)
\item qname: the appreviated name of each question, e.g. q1round2, q14round2
\item topic: brief textual description of the topic of each question
\item min: lower bound for the experts answer
\item max: upper bound for the experts answer
\item mode: mode of the answer
\item weighting: how much confidence the expert has in their answer (1-100)
\item expert: numerical identifier of the expert (1-10)
}}
\usage{
nov2013
}
\description{
A dataset containing the summarised answers from 10 different experts on 14 different
questions pertaining to seasonal distribution of right whales (Questions 1-8) and
the movements of right whales in and around the mid-Atlantic region (Questions 9-14).
}
\keyword{datasets}

