% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramSum.R
\name{paramSum}
\alias{paramSum}
\title{paramSum - Creates faceted summary of the Two Elicitations}
\usage{
paramSum(present = TRUE)
}
\description{
paramSum - Creates faceted summary of the Two Elicitations
}
\examples{
paramSum()
}
\seealso{
This is based off the summary plots I made for the shiny summary after the expert elicitation in June of 2015. The goal was to show how things changed between the two different elicitations. Or at least to be able to show a summary of the two different elicitations given that they were of two different forms.
}

