% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priorList}
\alias{priorList}
\title{Squashed Reformatted Priors from Round Two of Expert Elicitation of Right Whale Movement - June 2015}
\format{A recursive list with 2 elements, each of which is an 8 element list}
\usage{
priorList
}
\description{
A recursive list that holds the reformatted priors from each individual
expert for each sex. In \code{allqs} I have 6 elements each of which
contains 8 experts. However, what I really want is to combine those 6 into
two elements - one for each gender. That is what this list provides.
}
\keyword{datasets}

