% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumstat}
\alias{sumstat}
\title{Summarised Answers from Round Two of Expert Elicitation on Right Whale Movement - June 2015}
\format{A data frame with 6 rows and 4 variables:
\itemize{
\item confRawMean: mean of all 8 experts' answers for raw confidence (1-100)
\item confScaleMean: mean of all 8 experts' answers for scaled confidence (1-622)
\item moveMean: mean movements of number of whales out of 100
\item qname: the expert's name
}}
\usage{
sumstat
}
\description{
A dataset containing the summarised answers from 8 different experts on 6 different
questions pertaining to movements of right whales in and around the mid-Atlantic
region. Three questions were on Adult male right whales; 3 questions were on Adult
female right whales. List elements correspond to each question.
}
\keyword{datasets}

