% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcC.R
\name{transformC}
\alias{transformC}
\title{Tranform c matrix and return exponentiated values}
\usage{
transformC(cmat, tval = -600)
}
\description{
This function takes in the c matrix from the \code{calcC}
function, checks to make sure the values aren't below a
threshold (default value -600), and returns the exponentiated
values. If the values are are below -600, then it adds a scalar.
This scalar preserves the relationship between the values, which
represent the prior beliefs of individual experts
}

