% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winn}
\alias{winn}
\title{Historical Sightings Data from Winn et al. 1986}
\format{A data frame with 84 rows and 3 variables:
\itemize{
\item region: sighting region in which the whales were tallied (seus, mida, ne, gom, gsc, bof, rb)
\item month: month when whales were seen (Jan, Feb, Mar, Apr, May, Jun, Jul, Aug, Sep, Oct, Nov, Dec)
\item number: number of whales sighted (0, n)
}}
\usage{
winn
}
\description{
A dataset containing summary information taken from Figure 4 in Winn
et al. 1986. The variables are as follows:
}
\seealso{
The citation for this is Winn, et al. 1986. The Distributional Biology
of the Right Whale (\emph{Eubalaena glacialis}) in the Western North Atlantic.
Rep. Int. Whal. Comm. (Special Issue 10), pages 129-38.
Regions were taken from Figures 1 and 4. From Figure 4, I lumped Cape Hatteras, Mid-Atlantic,
and Long Island because these areas most closely approximate the regions from
NEAq's sightings catalog.
}
\keyword{datasets}

