% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/winnHist.R
\name{winnHist}
\alias{winnHist}
\title{winnHist - Makes a Faceted Histogram of Winn et al. Right Whale Distribution}
\usage{
winnHist(winn)
}
\arguments{
\item{winn}{Data frame of historical (pre-1986) distribution of right whales}
}
\description{
winnHist - Makes a Faceted Histogram of Winn et al. Right Whale Distribution
}
\examples{
winnHist(winn)
}
\seealso{
These data are from Winn, et al. 1986. The Distributional Biology of the Right Whale (Eubalaena glacialis) in the
Western North Atlantic. Rep. Int. Whal. Comm. (Special Issue 10), pages 129-38.
}

