% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makebdata.R
\name{makebdata}
\alias{makebdata}
\title{makebdata - Makes a long form of the data for plotting}
\usage{
makebdata(june2015)
}
\description{
makebdata - Makes a long form of the data for plotting
}
\examples{
makebdata(june2015)
}

